добавлено: transaction_schema.py(pydantic схема транзакций), transaction_service.py(логика работы транзакций), transaction_router.py(роутер /transaction) исправлено: float внутри AccountSchema заменен на Decimal тк были конфликты с Numeric из бд и float их схемы

+++ app/schemas/account_schema.py
@@ -1,8 +1,8 @@
 from pydantic import BaseModel, EmailStr, ConfigDict
-from sqlalchemy import Numeric
+from decimal import Decimal
 
 
 
 class AccountCreateSchema(BaseModel):
-    balance: float
+    balance: Decimal
